notebooks/salma-wahwah/requirements.txt
streamlit==1.19.0
google-generativeai==0.1.0
requests==2.28.1
